test(test): add aws mock server functions
