+++ htsget-search/Cargo.toml
@@ -11,7 +11,14 @@ homepage = "https://github.com/umccr/htsget-rs/blob/main/htsget-search/README.md
 repository = "https://github.com/umccr/htsget-rs"
 
 [features]
-s3-storage = ["dep:bytes", "dep:aws-sdk-s3", "dep:aws-config", "htsget-config/s3-storage", "htsget-test/s3-storage"]
+s3-storage = [
+    "dep:bytes",
+    "dep:aws-sdk-s3",
+    "dep:aws-config",
+    "htsget-config/s3-storage",
+    "htsget-test/s3-storage",
+    "htsget-test/aws-mocks"
+]
 default = ["s3-storage"]
 
 [dependencies]
@@ -51,12 +58,6 @@ serde = "1.0"
 [dev-dependencies]
 tempfile = "3.3"
 data-url = "0.2"
-
-# Aws S3 storage.
-aws-credential-types = { version = "0.54", features = ["test-util"] }
-s3s = { version = "0.3" }
-s3s-fs = { version = "0.3" }
-s3s-aws = { version = "0.3" }
 
 # Axum server
 reqwest = { version = "0.11", default-features = false, features = ["rustls-tls"] }
